+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TestTableRepairJob.java
@@ -165,7 +165,6 @@ public class TestTableRepairJob
 
         assertThat(myRepairJob.runnable()).isFalse();
 
-        verify(myRepairState, times(1)).update();
         verify(myRepairStateSnapshot, times(1)).canRepair();
     }
 
@@ -177,7 +176,6 @@ public class TestTableRepairJob
         mockRepairGroup(0L);
         assertThat(myRepairJob.runnable()).isTrue();
 
-        verify(myRepairState, times(1)).update();
         verify(myRepairStateSnapshot, times(2)).canRepair();
     }
 
@@ -190,7 +188,6 @@ public class TestTableRepairJob
         assertThat(myRepairJob.runnable()).isFalse();
         assertThat(myRepairJob.runnable()).isTrue();
 
-        verify(myRepairState, times(2)).update();
         verify(myRepairStateSnapshot, times(3)).canRepair();
     }