+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/RepairSchedulerImpl.java
@@ -185,8 +185,7 @@ public class RepairSchedulerImpl implements RepairScheduler, Closeable
                 .withRepairLockType(myRepairLockType)
                 .withRepairPolices(myRepairPolicies)
                 .build();
-
-        job.runnable();
+        job.refreshState();
 
         return job;
     }