+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TableRepairJob.java
@@ -103,7 +103,8 @@ public class TableRepairJob extends ScheduledJob
 
             for (ReplicaRepairGroup replicaRepairGroup : repairStateSnapshot.getRepairGroups())
             {
-                taskList.add(new RepairGroup(getRealPriority(), myTableReference, myRepairConfiguration,
+                taskList.add(new RepairGroup(getRealPriority(replicaRepairGroup.getLastCompletedAt()),
+                        myTableReference, myRepairConfiguration,
                         replicaRepairGroup, myJmxProxyFactory, myTableRepairMetrics,
                         myRepairLockType.getLockFactory(),
                         new RepairLockFactoryImpl(),
@@ -194,14 +195,6 @@ public class TableRepairJob extends ScheduledJob
     @Override
     public boolean runnable()
     {
-        try
-        {
-            myRepairState.update();
-        } catch (Exception e)
-        {
-            LOG.warn("Unable to check repair history, {}", this, e);
-        }
-
         RepairStateSnapshot repairStateSnapshot = myRepairState.getSnapshot();
 
         long lastRepaired = repairStateSnapshot.lastRepairedAt();
@@ -212,6 +205,19 @@ public class TableRepairJob extends ScheduledJob
         }
 
         return repairStateSnapshot.canRepair() && super.runnable();
+    }
+
+    @Override
+    public void refreshState()
+    {
+        try
+        {
+            myRepairState.update();
+        }
+        catch (Exception e)
+        {
+            LOG.warn("Unable to check repair history, {}", this, e);
+        }
     }
 
     /**