+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TestTableRepairJob.java
@@ -167,7 +167,6 @@ public class TestTableRepairJob
 
         assertThat(myRepairJob.runnable()).isFalse();
 
-        verify(myRepairState, times(1)).update();
         verify(myRepairStateSnapshot, times(1)).canRepair();
     }
 
@@ -179,7 +178,6 @@ public class TestTableRepairJob
         mockRepairGroup(0L);
         assertThat(myRepairJob.runnable()).isTrue();
 
-        verify(myRepairState, times(1)).update();
         verify(myRepairStateSnapshot, times(2)).canRepair();
     }
 
@@ -192,7 +190,6 @@ public class TestTableRepairJob
         assertThat(myRepairJob.runnable()).isFalse();
         assertThat(myRepairJob.runnable()).isTrue();
 
-        verify(myRepairState, times(2)).update();
         verify(myRepairStateSnapshot, times(3)).canRepair();
     }