+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Version 1.0.8 (Not yet released)
 
+* Fix priority calculation for local queue - Issue #546
 * Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
 * Fix malformed IPv6 for JMX - Issue #306