+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Changes
 
-## Version 4.0.5 (Not yet released)
+## Version 4.0.6 (Not yet released)
+
+## Version 4.0.5
 
 ### Merged from 1.0